% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_coefficients.R
\name{lr_coefficients}
\alias{lr_coefficients}
\title{lr_coefficients}
\usage{
lr_coefficients(predictor, covariates, data = NULL)
}
\arguments{
\item{predictor}{name of predictor "Y" variable}

\item{covariates}{vector of strings containing names of all the "X" covariates/parameters}

\item{data}{dataframe with data of interest}
}
\value{
Matrix of estimates of regression coefficient,
 standard errors, and t tests with p-values
}
\description{
performs single or multiple linear regression on selected predictor variable and covariates
}
\examples{
lr_coefficients("mpg", c("cyl", "disp", "hp"), data=mtcars)

}
