% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{mtcars}
\format{
## `mtcars`
A data frame with 32 rows and 11 columns:
\describe{
  \item{mpg}{miles per gallon}
  \item{cyl}{# of cylinders}
  \item{disp}{displacement, in cubic inches}
  \item{hp}{horsepower}
  \item{drat}{driveshaft ratio}
  \item{wt}{weight}
  \item{qsec}{1/4 mile time; a measure of acceleration}
  \item{vs}{'V' or straight - engine shape}
  \item{am}{1 = automatic transmission, 0 = manual transmission}
  \item{gear}{number of gears}
  \item{carb}{number of carburators}

}
}
\source{
"mtcars" originally came from the "Motor Trend Car Road Tests" in the 1974 Motor Trend US Magazine.
}
\usage{
mtcars
}
\description{
The dataset used for these examples is "mtcars".
It is built into base R but is included in this package to be extra sure that these examples will work.
}
\keyword{datasets}
